requirements.txt
certifi==2026.1.4
charset-normalizer==3.4.4
idna==3.11
pycryptodome==3.23.0
python-dotenv==1.2.1
requests==2.32.5
smile-id-core==3.0.1
types-requests==2.32.4.20260107
typing_extensions==4.15.0
urllib3==2.6.3
